+++ SmallBeachFarm/Framework/Config/ModConfig.cs
@@ -5,10 +5,10 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Framework.Config
     /// <summary>The mod configuration.</summary>
     internal class ModConfig
     {
-        /// <summary>Whether to enable the campfire.</summary>
+        /// <summary>Whether to add a functional campfire in front of the farmhouse.</summary>
         public bool AddCampfire { get; set; } = true;
 
-        /// <summary>Whether to use a map with islands for added space.</summary>
+        /// <summary>Whether to add ocean islands with extra land area.</summary>
         public bool EnableIslands { get; set; } = false;
 
         /// <summary>Use the beach's background music (i.e. wave sounds) on the beach farm.</summary>